refactor: apply classnames, change structure, fix some issues after review
