+++ .vscode/settings.json
@@ -5,6 +5,7 @@
     "Уолша",
     "Фрай",
     "Deciseconds",
+    "gendalf",
     "Pomodoro",
     "Reac",
     "Reactimer",
