+++ src/components/Total.js
